integrations/data-firewall/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.12.0
pydantic-settings==2.10.1
celery==5.5.3
redis==6.4.0
httpx==0.28.1
sqlalchemy==2.0.43
aiosqlite==0.21.0
structlog==25.4.0
slowapi==0.1.9
prometheus-client==0.23.1
aiodns>=3.5.0
pybreaker>=1.4.1
tenacity>=9.1.2
PyYAML>=6.0.2
pytest==8.4.2
pytest-asyncio==1.2.0
bandit==1.8.6
pip-audit==2.9.0
starlette>=0.49.1
